feat(dashboard): extraer ventana Descanso a módulo propio (#3736)
Extrae la ventana Modo descanso del monolito satellites.js a su propio
módulo views/dashboard/descanso.js, siguiendo el patrón de home.js /
multi-provider.js (shell inline, sin depender de pageShell del monolito
en demolición).

- NUEVO views/dashboard/descanso.js: renderDescanso (shell completo) +
  renderDescansoInner (fragmento embebible) + slug 'descanso'. Usa
  lib/escape-html.js (CA-B3) para la única interpolación SSR (opts.tz).
- satellites.js: renderModoDescanso queda como delegante de 1 línea
  (compat con HTML_ROUTES legacy y callers existentes).
- dashboard-routes.js: require defensivo de descansoView, guard en
  HTML_ROUTES['/modo-descanso'] y nuevo slug 'descanso' en VIEW_SLUGS.
  Path legacy /modo-descanso y slug ?view=descanso conviven sin redirect.
- dashboard.js: require defensivo de descansoView (patrón multi-provider).
- Tooltips (title + aria-label) en cada acción operativa (CA-C1):
  checkbox activar, guardar, + Periodo, ✕ eliminar.
- NUEVO test SSR + XSS guard (8 casos). Inventario V3 actualizado.

Closes #3736
